Add support for new insurance policy resource & endpoints (#139)
* Add support for new insurance policy resource & endpoints

This seems to be implemented as a separate service from the classic EHR and there are some
slight differences in the request & response structure.

First, the resource paths are nested under the /patients/{id} namespace, rather than having a
patient_id URL query parameter.

Second, updates are PUTs and require the full resource, rather the partial updates supported
by PATCH (and PUT) on most other endpoints.

Finally, pagination seems to be entirely different on the find endpoint, so left those new fields
out now until we can determine how pagination works, as there's no explanation of them in the docs
or how/if the page size, offset, etc. can be specified.

The POST endpoint for creates also supposedly allows for multiple policies, but left that out
for now as I don't think we have a particular use case for it right now.

* Split-out insurance policy update into separate struct